修正

+++ reducks/tweet/operation.ts
@@ -44,10 +44,12 @@ export const newPost = (Tweet: Tweet, navigation: any) => {
 }
 export const deletePost = (id: string | undefined,imageId: string | undefined) => {
     return async (dispatch: any) => {
-        await storage.ref('image').child(imageId + '.jpg').delete()
-        .then(()=>{
-            console.log("DELETE STORAGE")
-        })
+        if(imageId){
+            await storage.ref('image').child(imageId + '.jpg').delete()
+            .then(()=>{
+                console.log("DELETE STORAGE")
+            })
+        }
         postsRef.doc(id).delete()
             .then(() => {
                 dispatch(fetchTweets())
